+++ Source/Scene/Billboard.js
@@ -50,6 +50,8 @@ define([
      * updated, it may be more efficient to clear the collection with {@link BillboardCollection#removeAll}
      * and add new billboards instead of modifying each one.
      *
+     * @exception {DeveloperError} scaleByDistance.far must be greater than scaleByDistance.near
+     *
      * @see BillboardCollection
      * @see BillboardCollection#add
      * @see Label
@@ -77,18 +79,17 @@ define([
         this._alignedAxis = Cartesian3.clone(defaultValue(description.alignedAxis, Cartesian3.ZERO));
         this._width = description.width;
         this._height = description.height;
-        this._scaleByDistance = new NearFarScalar();
-        // disable scaleByDistance by default (scaling from 1.0 to 1.0)
-        //this.setScaleByDistance(defaultValue(description.scaleByDistance, new NearFarScalar(1.0, 1.0, 1.0e10, 1.0)));
-        // for initial testing with simple.czml and SandCastle CZML demo, using more interesting default value
-        // TODO: remove before final pull request into master
-        this.setScaleByDistance(defaultValue(description.scaleByDistance, new NearFarScalar(1.5e2, 1.5, 8.0e6, 0.0)));
+        this._scaleByDistance = description.scaleByDistance;
 
         this._pickId = undefined;
         this._pickIdThis = description._pickIdThis;
         this._billboardCollection = billboardCollection;
         this._dirty = false;
         this._index = -1; //Used only by BillboardCollection
+
+        if (defined(this._scaleByDistance) && this._scaleByDistance.far <= this._scaleByDistance.near) {
+            throw new DeveloperError('scaleByDistance.far must be greater than scaleByDistance.near.');
+        }
     };
 
     var SHOW_INDEX = Billboard.SHOW_INDEX = 0;
@@ -291,33 +292,49 @@ define([
      * Sets near and far scaling properties of a Billboard based on the billboard's distance from the camera.
      * A billboard's scale will interpolate between the {@link NearFarScalar#nearValue} and
      * {@link NearFarScalar#farValue} while the camera distance falls within the upper and lower bounds
-     * of the specified {@link NearFarScalar#nearDistance} and {@link NearFarScalar#farDistance}.
-     * Outside of these ranges the billboard's scale remains clamped to the nearest bound.
+     * of the specified {@link NearFarScalar#near} and {@link NearFarScalar#far}.
+     * Outside of these ranges the billboard's scale remains clamped to the nearest bound.  If undefined,
+     * scaleByDistance will be disabled.
      *
      * @memberof Billboard
      *
      * @param {NearFarScalar} scale The configuration of near and far distances and their respective scale values
      *
+     * @exception {DeveloperError} far distance must be greater than near distance.
+     *
      * @see Billboard#getScaleByDistance
      *
      * @example
+     * // Example 1.
      * // Set a billboard's scaleByDistance to scale by 1.5 when the
      * // camera is 1500 meters from the billboard and disappear as
      * // the camera distance approaches 8.0e6 meters.
      * b.setScaleByDistance(new NearFarScalar(1.5e2, 1.5, 8.0e6, 0.0));
+     *
+     * // Example 2.
+     * // disable scaling by distance
+     * b.setScaleByDistance(undefined);
      */
     Billboard.prototype.setScaleByDistance = function(scale) {
-        if (!defined(scale)) {
-            throw new DeveloperError('Requires NearFarScalar to configure the Billboard scale at near and far distances.');
+        if (defined(this._scaleByDistance) && defined(scale) && this._scaleByDistance.equals(scale)) {
+            return;
         }
 
-        if (scale.farDistance <= scale.nearDistance) {
-            throw new DeveloperError('farDistance must be greater than nearDistance and these distances cannot be equal.');
+        makeDirty(this, SCALE_BY_DISTANCE_INDEX);
+        if (!defined(scale)) {
+            this._scaleByDistance = undefined;
+            return;
+        }
+
+        if (scale.far <= scale.near) {
+            throw new DeveloperError('far distance must be greater than near distance.');
+        }
+
+        if (!defined(this._scaleByDistance)) {
+            this._scaleByDistance = new NearFarScalar();
         }
 
         NearFarScalar.clone(scale, this._scaleByDistance);
-
-        makeDirty(this, SCALE_BY_DISTANCE_INDEX);
     };
 
     /**
@@ -875,7 +892,8 @@ define([
                Cartesian3.equals(this._position, other._position) &&
                Color.equals(this._color, other._color) &&
                Cartesian2.equals(this._pixelOffset, other._pixelOffset) &&
-               Cartesian3.equals(this._eyeOffset, other._eyeOffset);
+               Cartesian3.equals(this._eyeOffset, other._eyeOffset) &&
+               NearFarScalar.equals(this._scaleByDistance, other._scaleByDistance);
     };
 
     Billboard.prototype._destroy = function() {